+++ src/charts/bar.js
@@ -12,7 +12,7 @@ define(function(require) {
     const d3Transition = require('d3-transition');
 
     const textHelper = require('./helpers/text');
-    const {exportChart} = require('./helpers/exportChart');
+    const {exportChart} = require('./../helpers/export');
     const colorHelper = require('../helpers/color');
     const {bar} = require('./helpers/loadingStates');
