+++ src/charts/sparkline.js
@@ -8,7 +8,7 @@ define(function(require){
     const d3Selection = require('d3-selection');
     const d3Transition = require('d3-transition');
 
-    const {exportChart} = require('./helpers/exportChart');
+    const {exportChart} = require('./../helpers/export');
     const colorHelper = require('../helpers/color');
     const {line} = require('./helpers/loadingStates');
     const {uniqueId} = require('./helpers/common');
